move fix logic to _common module
Since we already have a common module, it is better to put
the base fix logic inside that module instead of the fix module
itself
