+++ uaclient/api/u/pro/security/fix/_common/execute/v1.py
@@ -1,8 +1,7 @@
 from typing import List, Optional
 
 from uaclient import apt, messages, util
-from uaclient.api.u.pro.security.fix import FixStatus
-from uaclient.api.u.pro.security.fix._common import status_message
+from uaclient.api.u.pro.security.fix._common import FixStatus, status_message
 from uaclient.api.u.pro.security.fix._common.plan.v1 import (
     FixPlanAptUpgradeStep,
     FixPlanAttachStep,
