+++ uaclient/api/u/pro/security/fix/cve/execute/v1.py
@@ -2,8 +2,10 @@ from typing import List
 
 from uaclient.api.api import APIEndpoint
 from uaclient.api.data_types import AdditionalInfo
-from uaclient.api.u.pro.security.fix import FixStatus
-from uaclient.api.u.pro.security.fix._common import get_expected_overall_status
+from uaclient.api.u.pro.security.fix._common import (
+    FixStatus,
+    get_expected_overall_status,
+)
 from uaclient.api.u.pro.security.fix._common.execute.v1 import (
     FixExecuteResult,
     _execute_fix,
